+++ packages/did-ui-react/src/utils/socialLoginAuth.ts
@@ -1,8 +1,9 @@
 import { getCustomNetworkType, getServiceUrl } from '../components/config-provider/utils';
-import { PORTKEY_VERSION, WEB_PAGE, WEB_PAGE_TEST } from '../constants';
+import { WEB_PAGE, WEB_PAGE_TEST } from '../constants';
 import { ISocialLogin, NetworkType } from '../types';
 import { stringify } from 'query-string';
 import { dealURLLastChar } from './lib';
+import { facebookAuthPath, twitterAuthPath } from '../constants/network';
 
 export const socialLoginAuth = ({
   type,
@@ -23,8 +24,20 @@ export const socialLoginAuth = ({
   new Promise((resolve, reject) => {
     let timer: any = null;
     let serviceURI = dealURLLastChar(serviceUrl);
+    let _redirectURI = redirectURI;
 
-    if (type === 'Telegram' && !serviceURI) serviceURI = getServiceUrl();
+    if ((type === 'Telegram' || type === 'Facebook' || type === 'Twitter') && !serviceURI) serviceURI = getServiceUrl();
+    if (!redirectURI) {
+      switch (type) {
+        case 'Facebook':
+          _redirectURI = `${serviceURI}${facebookAuthPath}`;
+          break;
+        case 'Twitter':
+          _redirectURI = `${serviceURI}${twitterAuthPath}`;
+          break;
+      }
+    }
+
     const ctw = getCustomNetworkType();
 
     const thirdPage = ctw === 'Offline' ? WEB_PAGE_TEST : WEB_PAGE;
@@ -49,6 +62,7 @@ export const socialLoginAuth = ({
 
     window.addEventListener('message', onMessage);
     const baseUrl = `${thirdPage}/social-login/${type}`;
+    const encode = !(type === 'Twitter' || type === 'Facebook');
     const queryParams =
       type === 'Telegram'
         ? {
@@ -56,8 +70,14 @@ export const socialLoginAuth = ({
             from: 'openlogin',
             serviceURI,
           }
-        : { clientId, redirectURI, version: PORTKEY_VERSION };
-    const windowOpener = window.open(`${baseUrl}?${stringify(queryParams)}`);
+        : {
+            clientId,
+            redirectURI: _redirectURI,
+            // version: PORTKEY_VERSION
+          };
+
+    console.log(`${baseUrl}?${stringify(queryParams, { encode })}`, '=====baseUrl');
+    const windowOpener = window.open(`${baseUrl}?${stringify(queryParams, { encode })}`);
 
     timer = setInterval(() => {
       if (windowOpener?.closed) {